backend/requirements.txt
asgiref==3.9.1
colorama==0.4.6
Django==5.2.4
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
PyJWT==2.10.1
pyotp==2.9.0
python-dotenv==1.1.1
qrcode==8.2
sqlparse==0.5.3
tzdata==2025.2
